+++ CHANGELOG.md
@@ -17,6 +17,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Add dot-space-link-style rule [#165](https://github.com/nbbrd/heylogs/issues/165),[#323](https://github.com/nbbrd/heylogs/issues/323)
 - Add tagging SPI [#431](https://github.com/nbbrd/heylogs/issues/431)
 - Add tag versioning rule [#432](https://github.com/nbbrd/heylogs/issues/432)
+- Add domain configuration [#433](https://github.com/nbbrd/heylogs/issues/433)
 
 ### Changed